(PC-37028)[PRO] refactor: merge SummaryScreen inte tests
