+++ pro/src/components/IndividualOffer/SummaryScreen/SummaryScreen.spec.tsx
@@ -1,10 +1,21 @@
-import { act, fireEvent, screen } from '@testing-library/react'
+import { act, fireEvent, screen, waitFor, within } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { add, format, set, sub } from 'date-fns'
-import { Route, Routes } from 'react-router'
-import { vi } from 'vitest'
+import { add, addDays, format, set, sub } from 'date-fns'
+import { generatePath, Route, Routes } from 'react-router'
+import { expect } from 'vitest'
 
 import { api } from 'apiClient/api'
+import {
+  ApiError,
+  CancelablePromise,
+  GetIndividualOfferResponseModel,
+  GetMusicTypesResponse,
+  OfferStatus,
+  SubcategoryIdEnum,
+} from 'apiClient/v1'
+import { ApiRequestOptions } from 'apiClient/v1/core/ApiRequestOptions'
+import { ApiResult } from 'apiClient/v1/core/ApiResult'
+import * as useAnalytics from 'app/App/analytics/firebase'
 import {
   IndividualOfferContext,
   IndividualOfferContextValues,
@@ -14,34 +25,38 @@ import {
   OFFER_WIZARD_MODE,
 } from 'commons/core/Offers/constants'
 import { getIndividualOfferPath } from 'commons/core/Offers/utils/getIndividualOfferUrl'
+import { FORMAT_ISO_DATE_ONLY } from 'commons/utils/date'
+import { getAddressResponseIsLinkedToVenueModelFactory } from 'commons/utils/factories/commonOffersApiFactories'
 import {
   categoryFactory,
+  defaultGetOffererResponseModel,
   getIndividualOfferFactory,
+  getOfferManagingOffererFactory,
+  getOfferVenueFactory,
   individualOfferContextValuesFactory,
   subcategoryFactory,
+  venueListItemFactory,
 } from 'commons/utils/factories/individualApiFactories'
 import {
   currentOffererFactory,
   sharedCurrentUserFactory,
 } from 'commons/utils/factories/storeFactories'
-import { renderWithProviders } from 'commons/utils/renderWithProviders'
+import {
+  renderWithProviders,
+  RenderWithProvidersOptions,
+} from 'commons/utils/renderWithProviders'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
+import { Notification } from 'components/Notification/Notification'
 
 import { SummaryScreen } from './SummaryScreen'
 
-vi.mock('apiClient/api', () => ({
-  api: {
-    getOfferer: vi.fn(),
-    patchPublishOffer: vi.fn(),
-  },
-}))
-vi.mock('commons/utils/windowMatchMedia', () => ({
-  doesUserPreferReducedMotion: vi.fn(() => true),
-}))
-vi.mock('use-debounce', async () => ({
-  ...(await vi.importActual('use-debounce')),
-  useDebouncedCallback: vi.fn((fn) => fn),
-}))
+// vi.mock('apiClient/api', () => ({
+//   api: {
+//     getMusicTypes: vi.fn(),
+//     getOfferer: vi.fn(),
+//     patchPublishOffer: vi.fn(),
+//   },
+// }))
 
 const LABELS = {
   publicationModeNowRadio: /Publier maintenant/,
@@ -60,46 +75,833 @@ const ERROR_MESSAGES = {
     /Veuillez indiquer une date dans les 2 ans à venir/,
 }
 
-const renderSummaryScreen = ({
-  contextValue,
-  mode = OFFER_WIZARD_MODE.CREATION,
-  path,
-}: {
-  contextValue: IndividualOfferContextValues
-  mode?: OFFER_WIZARD_MODE
-  path?: string
-}) => {
-  const controlledPath =
-    path ??
-    getIndividualOfferPath({
-      step: OFFER_WIZARD_STEP_IDS.SUMMARY,
-      mode,
-    })
-  const element = (
-    <IndividualOfferContext.Provider value={contextValue}>
-      <SummaryScreen />
-    </IndividualOfferContext.Provider>
-  )
-  const component = (
-    <Routes>
-      <Route path={controlledPath} element={element} />
-    </Routes>
-  )
+const mockLogEvent = vi.fn()
 
-  const overrides = {
+const renderSummaryScreen = ({
+  contextValue = {},
+  mode = OFFER_WIZARD_MODE.READ_ONLY,
+  options,
+  path = getIndividualOfferPath({
+    step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+    mode,
+  }),
+}: Partial<{
+  contextValue: Partial<IndividualOfferContextValues>
+  mode: OFFER_WIZARD_MODE
+  options: RenderWithProvidersOptions
+  path: string
+  url: string
+}>) => {
+  const controlledContextValue =
+    individualOfferContextValuesFactory(contextValue)
+
+  return renderWithProviders(
+    <>
+      <IndividualOfferContext.Provider value={controlledContextValue}>
+        <Routes>
+          <Route
+            path={getIndividualOfferPath({
+              step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+              mode: OFFER_WIZARD_MODE.READ_ONLY,
+            })}
+            element={<SummaryScreen />}
+          />
+          <Route
+            path={getIndividualOfferPath({
+              step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+              mode: OFFER_WIZARD_MODE.CREATION,
+            })}
+            element={<SummaryScreen />}
+          />
+          <Route
+            path={getIndividualOfferPath({
+              step: OFFER_WIZARD_STEP_IDS.CONFIRMATION,
+              mode: OFFER_WIZARD_MODE.CREATION,
+            })}
+            element={<div>Confirmation page: creation</div>}
+          />
+        </Routes>
+      </IndividualOfferContext.Provider>
+      <Notification />
+    </>,
+    { initialRouterEntries: [path], ...options }
+  )
+}
+
+const categories = [categoryFactory({ id: 'A' })]
+
+const subCategories = [
+  subcategoryFactory({
+    id: String(SubcategoryIdEnum.CONCERT),
+    categoryId: 'A',
+  }),
+  subcategoryFactory({ categoryId: 'A' }),
+]
+
+describe('SummaryScreen', () => {
+  let customContext: Partial<IndividualOfferContextValues>
+  let musicTypes: GetMusicTypesResponse
+  beforeEach(() => {
+    musicTypes = [
+      {
+        gtl_id: '07000000',
+        label: 'Metal',
+        canBeEvent: true,
+      },
+      {
+        gtl_id: '02000000',
+        label: 'JAZZ / BLUES',
+        canBeEvent: true,
+      },
+      {
+        gtl_id: '03000000',
+        label: 'Bandes Originales',
+        canBeEvent: false,
+      },
+    ]
+    customContext = {
+      offer: getIndividualOfferFactory({
+        isEvent: false,
+        name: 'mon offre',
+        lastProvider: {
+          name: 'Ciné Office',
+        },
+        motorDisabilityCompliant: false,
+        mentalDisabilityCompliant: false,
+        audioDisabilityCompliant: false,
+        visualDisabilityCompliant: false,
+        description: 'ma description',
+        subcategoryId: SubcategoryIdEnum.CONCERT,
+        url: 'https://offer-url.example.com',
+        isDigital: true,
+        withdrawalDetails: 'détails de retrait',
+        bookingEmail: 'booking@example.com',
+        venue: getOfferVenueFactory({
+          name: 'ma venue',
+          publicName: 'ma venue (nom public)',
+          isVirtual: true,
+          managingOfferer: getOfferManagingOffererFactory({
+            name: 'mon offerer',
+          }),
+        }),
+      }),
+      subCategories,
+      categories,
+    }
+
+    vi.spyOn(useAnalytics, 'useAnalytics').mockImplementation(() => ({
+      logEvent: mockLogEvent,
+    }))
+
+    vi.spyOn(api, 'patchPublishOffer').mockResolvedValue(
+      getIndividualOfferFactory()
+    )
+    vi.spyOn(api, 'getMusicTypes').mockResolvedValue(musicTypes)
+  })
+
+  const expectOfferFields = async () => {
+    expect(await screen.findByText('Détails de l’offre')).toBeInTheDocument()
+    expect(screen.getByText('Type d’offre')).toBeInTheDocument()
+    expect(screen.getByText('Informations pratiques')).toBeInTheDocument()
+    expect(screen.getByText('Modalités d’accessibilité')).toBeInTheDocument()
+    expect(
+      screen.getByText('Notifications des réservations')
+    ).toBeInTheDocument()
+    expect(screen.getByText('Aperçu dans l’app')).toBeInTheDocument()
+
+    expect(screen.getByText(categories[0].proLabel)).toBeInTheDocument()
+    expect(screen.getByText(subCategories[0].proLabel)).toBeInTheDocument()
+    expect(screen.getByText('détails de retrait')).toBeInTheDocument()
+    expect(screen.getByText('Non accessible')).toBeInTheDocument()
+    expect(screen.getByText('booking@example.com')).toBeInTheDocument()
+    expect(screen.getAllByText('mon offre')).toHaveLength(2)
+    expect(screen.getAllByText('ma description')).toHaveLength(2)
+  }
+
+  it('should render component with informations on edition', async () => {
+    renderSummaryScreen({ contextValue: customContext })
+
+    await expectOfferFields()
+    expect(screen.getByText('Retour à la liste des offres')).toBeInTheDocument()
+    expect(screen.getByText('Visualiser dans l’app')).toBeInTheDocument()
+  })
+
+  it('should render public name if available', async () => {
+    renderSummaryScreen({ contextValue: customContext })
+
+    await expectOfferFields()
+    expect(screen.getByText('ma venue (nom public)')).toBeInTheDocument()
+  })
+
+  it('should render name if no public name available', async () => {
+    renderSummaryScreen({
+      contextValue: {
+        ...customContext,
+        offer: getIndividualOfferFactory({
+          ...customContext.offer,
+          venue: getOfferVenueFactory({
+            name: 'ma venue (name)',
+            publicName: null,
+          }),
+        }),
+      },
+    })
+
+    await expectOfferFields()
+    expect(screen.getByText('ma venue (name)')).toBeInTheDocument()
+  })
+
+  describe('On Creation', () => {
+    beforeEach(() => {
+      if (customContext.offer) {
+        customContext.offer = {
+          ...customContext.offer,
+          status: OfferStatus.DRAFT,
+        }
+      }
+    })
+
+    it('should render component with informations', async () => {
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      await expectOfferFields()
+      expect(
+        screen.queryByText('Visualiser dans l’app')
+      ).not.toBeInTheDocument()
+    })
+
+    it('should render component with new sections', async () => {
+      vi.spyOn(api, 'getOfferer').mockResolvedValue(
+        defaultGetOffererResponseModel
+      )
+      customContext.offer = getIndividualOfferFactory({ isEvent: true })
+
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      expect(
+        await screen.findByText('À propos de votre offre')
+      ).toBeInTheDocument()
+    })
+
+    it('should render component with right buttons', () => {
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      expect(screen.getByText('Retour')).toBeInTheDocument()
+      expect(
+        screen.getByText('Sauvegarder le brouillon et quitter')
+      ).toBeInTheDocument()
+      expect(screen.getByText('Publier l’offre')).toBeInTheDocument()
+    })
+
+    it('should link to creation confirmation page', async () => {
+      vi.spyOn(api, 'getOfferer').mockResolvedValue(
+        defaultGetOffererResponseModel
+      )
+
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      await userEvent.click(
+        screen.getByRole('button', { name: /Publier l’offre/ })
+      )
+
+      expect(
+        await screen.findByText(/Confirmation page: creation/)
+      ).toBeInTheDocument()
+    })
+
+    it('should disabled publish button link during submit', async () => {
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      const pageTitle = await screen.findByRole('heading', {
+        name: /Détails de l’offre/,
+      })
+      const buttonPublish = screen.getByRole('button', {
+        name: /Publier l’offre/,
+      })
+      expect(buttonPublish).not.toBeDisabled()
+
+      const mockResponse =
+        new CancelablePromise<GetIndividualOfferResponseModel>((resolve) =>
+          setTimeout(() => {
+            resolve(getIndividualOfferFactory())
+          }, 200)
+        )
+      vi.spyOn(api, 'patchPublishOffer').mockImplementationOnce(
+        () => mockResponse
+      )
+      vi.spyOn(api, 'getOfferer').mockResolvedValue(
+        defaultGetOffererResponseModel
+      )
+      await userEvent.click(buttonPublish)
+      expect(api.patchPublishOffer).toHaveBeenCalled()
+      expect(buttonPublish).toBeDisabled()
+      await waitFor(() => expect(pageTitle).not.toBeInTheDocument())
+      expect(
+        screen.getByText('Confirmation page: creation')
+      ).toBeInTheDocument()
+    })
+
+    it('should allow to publish offer later', async () => {
+      // Mock current date to avoid DST issues
+      vi.setSystemTime(new Date('2024-01-15T12:00:00.000Z'))
+
+      vi.spyOn(api, 'getOfferer').mockResolvedValue(
+        defaultGetOffererResponseModel
+      )
+      customContext.offer = getIndividualOfferFactory()
+
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      await userEvent.click(screen.getByLabelText(/Publier plus tard/))
+
+      await userEvent.click(
+        screen.getByRole('button', { name: 'Programmer l’offre' })
+      )
+
+      expect(
+        await screen.findByText('Veuillez sélectionner une date de publication')
+      ).toBeInTheDocument()
+      expect(
+        screen.getByText('Veuillez sélectionner une heure de publication')
+      ).toBeInTheDocument()
+      expect(
+        screen.queryByText(/Confirmation page: creation/)
+      ).not.toBeInTheDocument()
+
+      const publicationDate = format(
+        addDays(new Date(), 1),
+        FORMAT_ISO_DATE_ONLY
+      )
+
+      await userEvent.type(screen.getByLabelText('Date *'), publicationDate)
+      await userEvent.selectOptions(screen.getByLabelText('Heure *'), '11:00')
+
+      await userEvent.click(
+        screen.getByRole('button', { name: 'Programmer l’offre' })
+      )
+
+      expect(
+        await screen.findByText(/Confirmation page: creation/)
+      ).toBeInTheDocument()
+      expect(api.patchPublishOffer).toHaveBeenCalledWith({
+        id: customContext.offer.id,
+        publicationDatetime: `${publicationDate}T10:00:00Z`,
+      })
+
+      // Clean up the mocked time
+      vi.useRealTimers()
+    })
+
+    it('should allow to maker offer bookable later if the FF WIP_REFACTO_FUTURE_OFFER is enabled', async () => {
+      // Mock current date to avoid DST issues
+      vi.setSystemTime(new Date('2024-01-15T12:00:00.000Z'))
+
+      vi.spyOn(api, 'getOfferer').mockResolvedValue(
+        defaultGetOffererResponseModel
+      )
+      customContext.offer = getIndividualOfferFactory()
+
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+        options: { features: ['WIP_REFACTO_FUTURE_OFFER'] },
+      })
+
+      await userEvent.click(
+        screen.getByLabelText(/Rendre réservable plus tard/)
+      )
+
+      await userEvent.click(
+        screen.getByRole('button', { name: 'Publier l’offre' })
+      )
+
+      expect(
+        await screen.findByText(
+          'Veuillez sélectionner une date de réservabilité'
+        )
+      ).toBeInTheDocument()
+      expect(
+        screen.getByText('Veuillez sélectionner une heure de réservabilité')
+      ).toBeInTheDocument()
+      expect(
+        screen.queryByText(/Confirmation page: creation/)
+      ).not.toBeInTheDocument()
+
+      const bookingAllowedDate = format(
+        addDays(new Date(), 1),
+        FORMAT_ISO_DATE_ONLY
+      )
+
+      await userEvent.type(screen.getByLabelText('Date *'), bookingAllowedDate)
+      await userEvent.selectOptions(screen.getByLabelText('Heure *'), '11:00')
+
+      await userEvent.click(
+        screen.getByRole('button', { name: 'Publier l’offre' })
+      )
+
+      expect(
+        await screen.findByText(/Confirmation page: creation/)
+      ).toBeInTheDocument()
+      expect(api.patchPublishOffer).toHaveBeenCalledWith({
+        id: customContext.offer.id,
+        bookingAllowedDatetime: `${bookingAllowedDate}T10:00:00Z`,
+      })
+
+      // Clean up the mocked time
+      vi.useRealTimers()
+    })
+
+    it('should display notification on api error', async () => {
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      vi.spyOn(api, 'patchPublishOffer').mockRejectedValue(
+        new ApiError(
+          {} as ApiRequestOptions,
+          {
+            status: 400,
+            body: {},
+          } as ApiResult,
+          ''
+        )
+      )
+
+      await userEvent.click(
+        screen.getByRole('button', {
+          name: /Publier l’offre/,
+        })
+      )
+      expect(
+        await screen.findByText('Une erreur s’est produite, veuillez réessayer')
+      ).toBeInTheDocument()
+    })
+
+    it('should display redirect modal if first offer', async () => {
+      const venueId = 1
+      const context = {
+        offer: getIndividualOfferFactory({
+          venue: getOfferVenueFactory({ id: venueId }),
+        }),
+        showVenuePopin: {
+          [venueId]: true,
+        },
+      }
+
+      renderSummaryScreen({
+        contextValue: context,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+        options: { features: ['WIP_ENABLE_NEW_OFFER_CREATION_JOURNEY'] },
+      })
+
+      await userEvent.click(
+        screen.getByRole('button', { name: /Publier l’offre/ })
+      )
+
+      expect(
+        await screen.findByText('Félicitations, vous avez créé votre offre !')
+      ).toBeInTheDocument()
+      expect(
+        screen.getByRole('link', {
+          name: 'Ajouter un compte bancaire',
+        })
+      ).toHaveAttribute(
+        'href',
+        `/remboursements/informations-bancaires?structure=${context.offer.venue.managingOfferer.id}`
+      )
+    })
+
+    it('should display redirect modal if first non free offer', async () => {
+      const context = {
+        offer: getIndividualOfferFactory(),
+        venueList: [venueListItemFactory()],
+      }
+
+      vi.spyOn(api, 'getOfferer').mockResolvedValue({
+        ...defaultGetOffererResponseModel,
+        hasValidBankAccount: false,
+        hasPendingBankAccount: false,
+        hasNonFreeOffer: false,
+      })
+
+      vi.spyOn(api, 'patchPublishOffer').mockResolvedValue(
+        getIndividualOfferFactory({ isNonFreeOffer: true })
+      )
+
+      renderSummaryScreen({
+        contextValue: context,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+        options: {
+          features: ['WIP_ENABLE_NEW_OFFER_CREATION_JOURNEY'],
+        },
+      })
+
+      await userEvent.click(
+        screen.getByRole('button', { name: /Publier l’offre/ })
+      )
+
+      expect(
+        await screen.findByText('Félicitations, vous avez créé votre offre !')
+      ).toBeInTheDocument()
+      expect(
+        screen.getByRole('link', { name: 'Ajouter un compte bancaire' })
+      ).toHaveAttribute(
+        'href',
+        `/remboursements/informations-bancaires?structure=${context.offer.venue.managingOfferer.id}`
+      )
+    })
+
+    it('should not display redirect modal if hasPendingBankAccount is true', async () => {
+      const context = {
+        offer: getIndividualOfferFactory(),
+        venueList: [venueListItemFactory()],
+      }
+
+      vi.spyOn(api, 'getOfferer').mockResolvedValue({
+        ...defaultGetOffererResponseModel,
+        hasValidBankAccount: true,
+        hasPendingBankAccount: false,
+      })
+
+      vi.spyOn(api, 'patchPublishOffer').mockResolvedValue(
+        getIndividualOfferFactory({ isNonFreeOffer: false })
+      )
+
+      renderSummaryScreen({
+        contextValue: context,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+        options: {
+          features: ['WIP_ENABLE_NEW_OFFER_CREATION_JOURNEY'],
+        },
+      })
+
+      await userEvent.click(
+        screen.getByRole('button', { name: /Publier l’offre/ })
+      )
+
+      expect(
+        screen.queryByText('Félicitations, vous avez créé votre offre !')
+      ).not.toBeInTheDocument()
+    })
+
+    it('should not display redirect modal if offer is free', async () => {
+      vi.spyOn(api, 'patchPublishOffer').mockResolvedValue(
+        getIndividualOfferFactory({ isNonFreeOffer: false })
+      )
+
+      const context = {
+        offer: getIndividualOfferFactory(),
+        venueList: [venueListItemFactory()],
+      }
+
+      renderSummaryScreen({
+        contextValue: context,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+        options: {
+          features: ['WIP_ENABLE_NEW_OFFER_CREATION_JOURNEY'],
+        },
+      })
+
+      await userEvent.click(
+        screen.getByRole('button', { name: /Publier l’offre/ })
+      )
+
+      expect(
+        screen.queryByText('Félicitations, vous avez créé votre offre !')
+      ).not.toBeInTheDocument()
+    })
+
+    it('should not display redirect modal if venue hasNonFreeOffers', async () => {
+      vi.spyOn(api, 'getOfferer').mockResolvedValue({
+        ...defaultGetOffererResponseModel,
+        hasNonFreeOffer: true,
+      })
+
+      const context = {
+        offer: getIndividualOfferFactory(),
+        venueList: [venueListItemFactory()],
+      }
+
+      renderSummaryScreen({
+        contextValue: context,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+        options: {
+          features: ['WIP_ENABLE_NEW_OFFER_CREATION_JOURNEY'],
+        },
+      })
+
+      await userEvent.click(
+        screen.getByRole('button', { name: /Publier l’offre/ })
+      )
+
+      expect(
+        screen.queryByText('Félicitations, vous avez créé votre offre !')
+      ).not.toBeInTheDocument()
+    })
+
+    it('should render component with new sections and right address data', async () => {
+      vi.spyOn(api, 'getOfferer').mockResolvedValue(
+        defaultGetOffererResponseModel
+      )
+      customContext.offer = getIndividualOfferFactory({
+        isEvent: true,
+        address: {
+          ...getAddressResponseIsLinkedToVenueModelFactory({
+            label: 'mon adresse',
+            city: 'ma ville',
+            street: 'ma street',
+            postalCode: '1',
+          }),
+        },
+      })
+
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      expect(await screen.findByText(/Structure/)).toBeInTheDocument()
+      expect(
+        await screen.findByText('Localisation de l’offre')
+      ).toBeInTheDocument()
+
+      // Present without the colon « : » in the <VenueDetails /> component
+      expect(screen.getByText('Intitulé')).toBeInTheDocument()
+      expect(screen.getByText('Adresse')).toBeInTheDocument()
+
+      // Present with the colon « : » in the <SummaryScreen /> component
+      expect(screen.getByText('Intitulé :')).toBeInTheDocument()
+      expect(screen.getByText('Adresse :')).toBeInTheDocument()
+
+      // Both present in <VenueDetails /> and <SummaryScreen /> components
+      expect(screen.getAllByText('mon adresse')).toHaveLength(2)
+      expect(screen.getAllByText('ma street 1 ma ville')).toHaveLength(2)
+    })
+
+    it('should render component with new sections and empty address data', async () => {
+      vi.spyOn(api, 'getOfferer').mockResolvedValue(
+        defaultGetOffererResponseModel
+      )
+      customContext.offer = getIndividualOfferFactory({
+        isEvent: true,
+        address: null,
+      })
+
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      expect(await screen.findByText(/Structure/)).toBeInTheDocument()
+      expect(
+        await screen.findByText('Localisation de l’offre')
+      ).toBeInTheDocument()
+
+      expect(
+        within(screen.getByTestId('localisation-offer-details')).getAllByText(
+          '-'
+        )
+      ).toHaveLength(2)
+    })
+  })
+
+  describe('banners', () => {
+    it('should display pre publishing banner in creation', () => {
+      const url = generatePath(
+        getIndividualOfferPath({
+          step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+          mode: OFFER_WIZARD_MODE.CREATION,
+        }),
+        { offerId: 'AA' }
+      )
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: url,
+      })
+
+      expect(screen.getByText('Vous y êtes presque !')).toBeInTheDocument()
+    })
+
+    it('should display a notification when saving as draft', async () => {
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.CREATION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.CREATION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+      await userEvent.click(
+        screen.getByText('Sauvegarder le brouillon et quitter')
+      )
+      expect(
+        screen.getByText('Brouillon sauvegardé dans la liste des offres')
+      ).toBeInTheDocument()
+    })
+
+    it('should not display pre publishing banner in edition mode', () => {
+      renderSummaryScreen({
+        contextValue: customContext,
+        mode: OFFER_WIZARD_MODE.EDITION,
+        path: generatePath(
+          getIndividualOfferPath({
+            step: OFFER_WIZARD_STEP_IDS.SUMMARY,
+            mode: OFFER_WIZARD_MODE.EDITION,
+          }),
+          { offerId: 'AA' }
+        ),
+      })
+
+      expect(
+        screen.queryByText('Vous y êtes presque !')
+      ).not.toBeInTheDocument()
+    })
+  })
+})
+
+describe('Form', () => {
+  const inOneMonth = set(add(new Date(), { months: 1 }), {
+    hours: 10,
+    minutes: 0,
+  })
+
+  let contextValue: IndividualOfferContextValues
+  const mode = OFFER_WIZARD_MODE.CREATION
+  const options: RenderWithProvidersOptions = {
     user: sharedCurrentUserFactory(),
-    initialRouterEntries: [controlledPath],
     storeOverrides: {
       user: { currentUser: sharedCurrentUserFactory() },
       offerer: currentOffererFactory(),
     },
   }
-
-  return renderWithProviders(component, overrides)
-}
-
-describe('SummaryScreen', () => {
-  let contextValue: IndividualOfferContextValues
 
   beforeEach(() => {
     const categories = [
@@ -135,187 +937,181 @@ describe('SummaryScreen', () => {
       offer: getIndividualOfferFactory({ id: 1 }),
       subCategories,
     })
+
+    vi.spyOn(api, 'getOfferer').mockImplementation(vi.fn())
+    vi.spyOn(api, 'patchPublishOffer').mockResolvedValue(
+      getIndividualOfferFactory()
+    )
   })
 
   afterEach(() => {
     vi.clearAllMocks()
   })
 
-  describe('Form', () => {
-    const inOneMonth = set(add(new Date(), { months: 1 }), {
+  it("should validate publication date and time when it's a scheduled publication", async () => {
+    renderSummaryScreen({ contextValue, mode, options })
+
+    await userEvent.click(
+      screen.getByLabelText(LABELS.publicationModeLaterRadio)
+    )
+    await userEvent.click(
+      screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
+    )
+
+    expect(
+      await screen.findByText(ERROR_MESSAGES.publicationDateIsRequired)
+    ).toBeVisible()
+    expect(
+      await screen.findByText(ERROR_MESSAGES.publicationTimeIsRequired)
+    ).toBeVisible()
+
+    const publicationDateInput = screen.getByLabelText(
+      LABELS.publicationDateInput
+    )
+    const publicationTimeSelect = screen.getByLabelText(
+      LABELS.publicationTimeSelect
+    )
+
+    await userEvent.type(publicationDateInput, format(inOneMonth, 'yyyy-MM-dd'))
+
+    expect(
+      screen.queryByText(ERROR_MESSAGES.publicationDateIsRequired)
+    ).not.toBeInTheDocument()
+    expect(
+      screen.queryByText(ERROR_MESSAGES.publicationDateMustBeInFuture)
+    ).toBeVisible()
+    expect(
+      screen.queryByText(ERROR_MESSAGES.publicationTimeIsRequired)
+    ).toBeVisible()
+
+    await userEvent.selectOptions(
+      publicationTimeSelect,
+      format(inOneMonth, 'HH:mm')
+    )
+
+    expect(
+      screen.queryByTestId('error-publicationDate')
+    ).not.toBeInTheDocument()
+
+    await userEvent.click(
+      screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
+    )
+
+    expect(api.patchPublishOffer).toHaveBeenCalledWith({
+      id: contextValue.offer!.id,
+      publicationDatetime: expect.any(String),
+      bookingAllowedDatetime: undefined,
+    })
+  })
+
+  it("should require publication date to be in the future when it's a scheduled publication", async () => {
+    const yesterday = set(sub(new Date(), { days: 1 }), {
       hours: 10,
       minutes: 0,
     })
 
-    it("should validate publication date and time when it's a scheduled publication", async () => {
-      renderSummaryScreen({ contextValue })
+    renderSummaryScreen({ contextValue, mode, options })
 
-      await userEvent.click(
-        screen.getByLabelText(LABELS.publicationModeLaterRadio)
-      )
-      await userEvent.click(
-        screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
-      )
+    await userEvent.click(
+      screen.getByLabelText(LABELS.publicationModeLaterRadio)
+    )
 
-      expect(
-        await screen.findByText(ERROR_MESSAGES.publicationDateIsRequired)
-      ).toBeVisible()
-      expect(
-        await screen.findByText(ERROR_MESSAGES.publicationTimeIsRequired)
-      ).toBeVisible()
+    const publicationDateInput = screen.getByLabelText(
+      LABELS.publicationDateInput
+    )
+    const publicationTimeSelect = screen.getByLabelText(
+      LABELS.publicationTimeSelect
+    )
 
-      const publicationDateInput = screen.getByLabelText(
-        LABELS.publicationDateInput
-      )
-      const publicationTimeSelect = screen.getByLabelText(
-        LABELS.publicationTimeSelect
-      )
+    await userEvent.type(publicationDateInput, format(yesterday, 'yyyy-MM-dd'))
+    await userEvent.selectOptions(
+      publicationTimeSelect,
+      format(yesterday, 'HH:mm')
+    )
 
-      await userEvent.type(
-        publicationDateInput,
-        format(inOneMonth, 'yyyy-MM-dd')
-      )
+    expect(
+      await screen.findByText(ERROR_MESSAGES.publicationDateMustBeInFuture)
+    ).toBeVisible()
 
-      expect(
-        screen.queryByText(ERROR_MESSAGES.publicationDateIsRequired)
-      ).not.toBeInTheDocument()
-      expect(
-        screen.queryByText(ERROR_MESSAGES.publicationDateMustBeInFuture)
-      ).toBeVisible()
-      expect(
-        screen.queryByText(ERROR_MESSAGES.publicationTimeIsRequired)
-      ).toBeVisible()
+    // await act(() =>
+    //   fireEvent.input(publicationDateInput, {
+    //     target: {
+    //       value: format(inOneMonth, 'yyyy-MM-dd'),
+    //     },
+    //   })
+    // )
+    await userEvent.clear(publicationDateInput)
+    await userEvent.type(publicationDateInput, format(inOneMonth, 'yyyy-MM-dd'))
 
-      await userEvent.selectOptions(
-        publicationTimeSelect,
-        format(inOneMonth, 'HH:mm')
-      )
+    expect(
+      screen.queryByTestId('error-publicationDate')
+    ).not.toBeInTheDocument()
 
-      expect(
-        screen.queryByTestId('error-publicationDate')
-      ).not.toBeInTheDocument()
+    await userEvent.click(
+      screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
+    )
 
-      await userEvent.click(
-        screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
-      )
+    expect(api.patchPublishOffer).toHaveBeenCalledWith({
+      id: contextValue.offer!.id,
+      publicationDatetime: expect.any(String),
+      bookingAllowedDatetime: undefined,
+    })
+  })
 
-      expect(api.patchPublishOffer).toHaveBeenCalledWith({
-        id: contextValue.offer!.id,
-        publicationDatetime: expect.any(String),
-        bookingAllowedDatetime: undefined,
-      })
+  it("should require publication date to be within two years when it's a scheduled publication", async () => {
+    const inMoreThanTwoYears = set(add(new Date(), { months: 25 }), {
+      hours: 10,
+      minutes: 0,
     })
 
-    it("should require publication date to be in the future when it's a scheduled publication", async () => {
-      const yesterday = set(sub(new Date(), { days: 1 }), {
-        hours: 10,
-        minutes: 0,
+    renderSummaryScreen({ contextValue, mode, options })
+
+    await userEvent.click(
+      screen.getByLabelText(LABELS.publicationModeLaterRadio)
+    )
+
+    const publicationDateInput = screen.getByLabelText(
+      LABELS.publicationDateInput
+    )
+    const publicationTimeSelect = screen.getByLabelText(
+      LABELS.publicationTimeSelect
+    )
+
+    await userEvent.type(
+      publicationDateInput,
+      format(inMoreThanTwoYears, 'yyyy-MM-dd')
+    )
+    await userEvent.selectOptions(
+      publicationTimeSelect,
+      format(inMoreThanTwoYears, 'HH:mm')
+    )
+
+    expect(
+      await screen.findByText(
+        ERROR_MESSAGES.publicationDateMustBeWithinTwoYears
+      )
+    ).toBeVisible()
+
+    await act(() =>
+      fireEvent.input(publicationDateInput, {
+        target: {
+          value: format(inOneMonth, 'yyyy-MM-dd'),
+        },
       })
+    )
 
-      renderSummaryScreen({ contextValue })
+    expect(
+      screen.queryByTestId('error-publicationDate')
+    ).not.toBeInTheDocument()
 
-      await userEvent.click(
-        screen.getByLabelText(LABELS.publicationModeLaterRadio)
-      )
+    await userEvent.click(
+      screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
+    )
 
-      const publicationDateInput = screen.getByLabelText(
-        LABELS.publicationDateInput
-      )
-      const publicationTimeSelect = screen.getByLabelText(
-        LABELS.publicationTimeSelect
-      )
-
-      await userEvent.type(
-        publicationDateInput,
-        format(yesterday, 'yyyy-MM-dd')
-      )
-      await userEvent.selectOptions(
-        publicationTimeSelect,
-        format(yesterday, 'HH:mm')
-      )
-
-      expect(
-        await screen.findByText(ERROR_MESSAGES.publicationDateMustBeInFuture)
-      ).toBeVisible()
-
-      await act(() =>
-        fireEvent.input(publicationDateInput, {
-          target: {
-            value: format(inOneMonth, 'yyyy-MM-dd'),
-          },
-        })
-      )
-
-      expect(
-        screen.queryByTestId('error-publicationDate')
-      ).not.toBeInTheDocument()
-
-      await userEvent.click(
-        screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
-      )
-
-      expect(api.patchPublishOffer).toHaveBeenCalledWith({
-        id: contextValue.offer!.id,
-        publicationDatetime: expect.any(String),
-        bookingAllowedDatetime: undefined,
-      })
-    })
-
-    it("should require publication date to be within two years when it's a scheduled publication", async () => {
-      const inMoreThanTwoYears = set(add(new Date(), { months: 25 }), {
-        hours: 10,
-        minutes: 0,
-      })
-
-      renderSummaryScreen({ contextValue })
-
-      await userEvent.click(
-        screen.getByLabelText(LABELS.publicationModeLaterRadio)
-      )
-
-      const publicationDateInput = screen.getByLabelText(
-        LABELS.publicationDateInput
-      )
-      const publicationTimeSelect = screen.getByLabelText(
-        LABELS.publicationTimeSelect
-      )
-
-      await userEvent.type(
-        publicationDateInput,
-        format(inMoreThanTwoYears, 'yyyy-MM-dd')
-      )
-      await userEvent.selectOptions(
-        publicationTimeSelect,
-        format(inMoreThanTwoYears, 'HH:mm')
-      )
-
-      expect(
-        await screen.findByText(
-          ERROR_MESSAGES.publicationDateMustBeWithinTwoYears
-        )
-      ).toBeVisible()
-
-      await act(() =>
-        fireEvent.input(publicationDateInput, {
-          target: {
-            value: format(inOneMonth, 'yyyy-MM-dd'),
-          },
-        })
-      )
-
-      expect(
-        screen.queryByTestId('error-publicationDate')
-      ).not.toBeInTheDocument()
-
-      await userEvent.click(
-        screen.getByRole('button', { name: LABELS.submitScheduledOfferButton })
-      )
-
-      expect(api.patchPublishOffer).toHaveBeenCalledWith({
-        id: contextValue.offer!.id,
-        publicationDatetime: expect.any(String),
-        bookingAllowedDatetime: undefined,
-      })
+    expect(api.patchPublishOffer).toHaveBeenCalledWith({
+      id: contextValue.offer!.id,
+      publicationDatetime: expect.any(String),
+      bookingAllowedDatetime: undefined,
     })
   })
 })
